🍪😀 Checkpoint ./config/config.json:41556106/151 ./events/ready.js:41556106/535 ./commands/dev/stats.js:41556106/11 ./commands/public/top.js:41556106/102 ./commands/admin/whitelist.js:41556106/57

+++ commands/admin/whitelist.js
@@ -68,7 +68,7 @@ module.exports.run = async(client, message, args) => {
       });
 
     }
-    if (args[1].toLowerCase() === "remove") {
+    if (args[1] && args[1].toLowerCase() === "remove") {
      const userData = await ownersSchema.findOne({userId: user.id, guildId: message.guild.id})
      if (!userData) return message.replyNoMention(`**هذا العضو ليس لديه اي تحكمات بالبوت بالفعل😊**`)  
      userData.delete();
